test(statistics): split ExplorerResultsTablePresenterTest by table behaviour
Separate columns, matrix, box-plot, and hospital-metric cases behind a shared presenter test case.
